+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Supplier
+- Supplier Invoice
 
 ## [0.2.0] - 2021-02-03
 ### Added